refactor: 결과 페이지에서 결과 화면 컴포넌트 분리
